refactor(web): move assignment duration helpers out of the route modules
main added a lint rule banning vitest imports under src/routes, which the two
tests this branch placed there now trip. `parseDurationDays` and
`getDefaultAssignmentExpiry` were only exported so those tests could reach
them, so move both to a util instead — the route files export just `Route`
again, and the min/max/default of one rule now live together.
